+++ packages/chakra-ui-nuxt/example/autoimport/nuxt.config.js
@@ -2,11 +2,11 @@ const { resolve } = require('path')
 const theme = require('./utils/theme')
 
 module.exports = {
-  rootDir: resolve(__dirname, '..'),
+  rootDir: resolve(__dirname, '../..'),
   buildDir: resolve(__dirname, '.nuxt'),
   srcDir: __dirname,
   modules: [
-    { handler: require('../lib/module') },
+    { handler: require('../../lib/module') },
     { handler: require('@nuxtjs/emotion') }
   ],
   chakra: {