+++ packages/chakra-ui-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 0.6.6
+
+### Patch Changes
+
+**🌟 New features**
+- feat: include all known css properties #319
+- feat(CStack): allow stack html elements #324
+
+**🛠 Bug fixes**
+- fix(creset): include font-family to global styles #323
+- fix(inputgroup): default inputgroup z-index to containing element zindex #325
+- Documentation fixes #329, #328, #321
+
 ## 0.6.5
 
 ### Patch Changes